// src/shared/lib/dataAccess/search/generated/model/getEvents200.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * Search API
 * With UiTdatabank's Search API you can search events, places and organizers.
 * OpenAPI spec version: 3.0
 */
import type { Event } from './event';
import type { GetEvents200Facet } from './getEvents200Facet';

export type GetEvents200 = {
  /** The amount of results that is being returned per page. */
  itemsPerPage: number;
  /** Total amount of results for the given query parameters. */
  totalItems: number;
  /** Search results (paginated). Note that the complete search results details will only be returned if `?embed=true` is used. Otherwise only the `@id` will be returned. */
  member: Event[];
  /** Facet counts per possible filter & value. */
  facet?: GetEvents200Facet;
};